+++ examples/rest-server/rest-authentication.h
@@ -1,0 +1,39 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2018 8devices
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#ifndef REST_AUTHENTICATION_H
+#define REST_AUTHENTICATION_H
+
+#include <ulfius.h>
+
+#define HEADER_AUTHORIZATION   "Authorization"
+#define HEADER_UNAUTHORIZED    "WWW-Authenticate"
+#define HEADER_PREFIX_BEARER   "Bearer " 
+
+int rest_authenticate_cb(const struct _u_request *request, struct _u_response *response,
+                         void *user_data);
+int rest_validate_jwt_cb(const struct _u_request *request, struct _u_response *response,
+                    void *user_data);
+
+#endif // REST_AUTHENTICATION_H
