+++ examples/rest-server/rest-core-types.h
@@ -66,6 +66,7 @@ typedef struct
     const char *name;
 } rest_notif_timeout_t;
 
+size_t rest_get_random(void *buf, size_t buflen);
 
 rest_async_response_t *rest_async_response_new(void);
